Fixed Playwright Test Failures & Completed Component Consolidation
