+++ components/sticky-notes-demo.tsx
@@ -3,48 +3,80 @@
 import * as React from "react"
 import { useState } from "react"
 import { Button } from "@/components/ui/button"
-import { StickyNoteCard } from "@/components/sticky-note-card"
-import type { DemoNote } from "@/lib/types"
+import { Note } from "@/components/note"
+import type { Note as NoteType, ChecklistItem } from "@/lib/types"
 import { Plus } from "lucide-react"
 import { motion, AnimatePresence } from "framer-motion"
 
-const initialNotes: DemoNote[] = [
+const initialNotes: NoteType[] = [
   {
-    id: 1,
-    author: { name: "Sahil", initial: "S" },
+    id: "1",
+    content: "",
     color: "bg-green-200/70",
-    tasks: [
-      { id: 101, text: "Gumboard release by Friday", completed: false },
-      { id: 102, text: "Finance update by Friday", completed: false },
-      { id: 103, text: "Jacquez", completed: true },
+    done: false,
+    createdAt: new Date().toISOString(),
+    updatedAt: new Date().toISOString(),
+    user: {
+      id: "user-1",
+      name: "Sahil",
+      email: "sahil@example.com",
+    },
+    checklistItems: [
+      { id: "101", content: "Gumboard release by Friday", checked: false, order: 0 },
+      { id: "102", content: "Finance update by Friday", checked: false, order: 1 },
+      { id: "103", content: "Jacquez", checked: true, order: 2 },
     ],
   },
   {
-    id: 2,
-    author: { name: "Michelle", initial: "M" },
+    id: "2",
+    content: "",
     color: "bg-purple-200/60",
-    tasks: [
-      { id: 201, text: "Helper Tix (Mon-Fri)", completed: false },
-      { id: 202, text: "Active Refunds (2x a week)", completed: false },
-      { id: 203, text: "Card Tester Metabase (DAILY)", completed: true },
+    done: false,
+    createdAt: new Date().toISOString(),
+    updatedAt: new Date().toISOString(),
+    user: {
+      id: "user-2",
+      name: "Michelle",
+      email: "michelle@example.com",
+    },
+    checklistItems: [
+      { id: "201", content: "Helper Tix (Mon-Fri)", checked: false, order: 0 },
+      { id: "202", content: "Active Refunds (2x a week)", checked: false, order: 1 },
+      { id: "203", content: "Card Tester Metabase (DAILY)", checked: true, order: 2 },
     ],
   },
   {
-    id: 3,
-    author: { name: "Steve", initial: "S" },
+    id: "3",
+    content: "",
     color: "bg-blue-200/60",
-    tasks: [
-      { id: 301, text: "Review support huddle", completed: false },
-      { id: 302, text: "Metabase queries", completed: false },
+    done: false,
+    createdAt: new Date().toISOString(),
+    updatedAt: new Date().toISOString(),
+    user: {
+      id: "user-3",
+      name: "Steve",
+      email: "steve@example.com",
+    },
+    checklistItems: [
+      { id: "301", content: "Review support huddle", checked: false, order: 0 },
+      { id: "302", content: "Metabase queries", checked: false, order: 1 },
     ],
   },
   {
-    id: 4,
-    author: { name: "Daniel", initial: "D" },
+    id: "4",
+    content: "",
     color: "bg-pink-200/70",
-    tasks: [
-      { id: 401, text: "Fixed unnecessary description", completed: false },
-      { id: 402, text: "PR reviews", completed: true },
+    done: false,
+    createdAt: new Date().toISOString(),
+    updatedAt: new Date().toISOString(),
+    user: {
+      id: "user-4",
+      name: "Daniel",
+      email: "daniel@example.com",
+    },
+    checklistItems: [
+      { id: "401", content: "Fixed unnecessary description", checked: false, order: 0 },
+      { id: "402", content: "PR reviews", checked: true, order: 1 },
     ],
   },
 ]
@@ -272,24 +304,123 @@ const itemVariants = {
 }
 
 export function StickyNotesDemo() {
-  const [notes, setNotes] = useState<DemoNote[]>(initialNotes)
+  const [notes, setNotes] = useState<NoteType[]>(initialNotes)
 
-  const handleUpdateNote = (updatedNote: DemoNote) => {
+  const handleUpdateNote = (updatedNote: NoteType) => {
     setNotes(notes.map((note) => (note.id === updatedNote.id ? updatedNote : note)))
   }
 
-  const handleDeleteNote = (noteId: number) => {
+  const handleDeleteNote = (noteId: string) => {
     setNotes(notes.filter((note) => note.id !== noteId))
+  }
+
+  const handleAddChecklistItem = (noteId: string, content: string) => {
+    const note = notes.find((n) => n.id === noteId)
+    if (!note) return
+
+    const newItem: ChecklistItem = {
+      id: `${Date.now()}`,
+      content,
+      checked: false,
+      order: (note.checklistItems || []).length,
+    }
+
+    const updatedNote = {
+      ...note,
+      checklistItems: [...(note.checklistItems || []), newItem],
+    }
+
+    handleUpdateNote(updatedNote)
+  }
+
+  const handleToggleChecklistItem = (noteId: string, itemId: string) => {
+    const note = notes.find((n) => n.id === noteId)
+    if (!note || !note.checklistItems) return
+
+    const updatedItems = note.checklistItems.map((item) =>
+      item.id === itemId ? { ...item, checked: !item.checked } : item
+    )
+
+    const allItemsChecked = updatedItems.every((item) => item.checked)
+
+    const updatedNote = {
+      ...note,
+      checklistItems: updatedItems,
+      done: allItemsChecked,
+    }
+
+    handleUpdateNote(updatedNote)
+  }
+
+  const handleEditChecklistItem = (noteId: string, itemId: string, content: string) => {
+    const note = notes.find((n) => n.id === noteId)
+    if (!note || !note.checklistItems) return
+
+    const updatedItems = note.checklistItems.map((item) =>
+      item.id === itemId ? { ...item, content } : item
+    )
+
+    const updatedNote = {
+      ...note,
+      checklistItems: updatedItems,
+    }
+
+    handleUpdateNote(updatedNote)
+  }
+
+  const handleDeleteChecklistItem = (noteId: string, itemId: string) => {
+    const note = notes.find((n) => n.id === noteId)
+    if (!note || !note.checklistItems) return
+
+    const updatedItems = note.checklistItems.filter((item) => item.id !== itemId)
+    const allItemsChecked = updatedItems.every((item) => item.checked)
+
+    const updatedNote = {
+      ...note,
+      checklistItems: updatedItems,
+      done: updatedItems.length > 0 ? allItemsChecked : false,
+    }
+
+    handleUpdateNote(updatedNote)
+  }
+
+  const handleToggleAllChecklistItems = (noteId: string) => {
+    const note = notes.find((n) => n.id === noteId)
+    if (!note || !note.checklistItems) return
+
+    const allChecked = note.checklistItems.every((item) => item.checked)
+    const updatedItems = note.checklistItems.map((item) => ({
+      ...item,
+      checked: !allChecked,
+    }))
+
+    const updatedNote = {
+      ...note,
+      checklistItems: updatedItems,
+      done: !allChecked,
+    }
+
+    handleUpdateNote(updatedNote)
   }
 
   const handleAddNote = () => {
     const randomColor = noteColors[Math.floor(Math.random() * noteColors.length)]
     const randomAuthor = authors[Math.floor(Math.random() * authors.length)]
-    const newNote: DemoNote = {
-      id: Date.now(),
-      author: randomAuthor,
+    const newNote: NoteType = {
+      id: `${Date.now()}`,
+      content: "",
       color: randomColor,
-      tasks: [{ id: Date.now() + 1, text: "New to-do", completed: false }],
+      done: false,
+      createdAt: new Date().toISOString(),
+      updatedAt: new Date().toISOString(),
+      user: {
+        id: `user-${Date.now()}`,
+        name: randomAuthor.name,
+        email: `${randomAuthor.name.toLowerCase().replace(/\s+/g, '')}@example.com`,
+      },
+      checklistItems: [
+        { id: `${Date.now() + 1}`, content: "New to-do", checked: false, order: 0 },
+      ],
     }
     setNotes([newNote, ...notes])
   }
@@ -312,7 +443,19 @@ export function StickyNotesDemo() {
           <AnimatePresence>
             {notes.map((note) => (
               <motion.div key={note.id} className="mb-4 break-inside-avoid" variants={itemVariants} exit="exit" layout>
-                <StickyNoteCard note={note} onUpdate={handleUpdateNote} onDelete={handleDeleteNote} />
+                <Note
+                  note={note}
+                  currentUser={{ id: "demo-user", name: "Demo User", email: "demo@example.com" }}
+                  onUpdate={handleUpdateNote}
+                  onDelete={handleDeleteNote}
+                  onAddChecklistItem={handleAddChecklistItem}
+                  onToggleChecklistItem={handleToggleChecklistItem}
+                  onEditChecklistItem={handleEditChecklistItem}
+                  onDeleteChecklistItem={handleDeleteChecklistItem}
+                  onToggleAllChecklistItems={handleToggleAllChecklistItems}
+                  className={`${note.color} bg-white dark:bg-zinc-900`}
+                  style={{ padding: "16px" }}
+                />
               </motion.div>
             ))}
           </AnimatePresence>
